chore(icons): move to viewer package
