+++ packages/form-js-viewer/src/render/components/index.js
@@ -42,3 +42,5 @@ export const formFields = [
   Textfield,
   Textarea
 ];
+
+export * from './icons';
